rename clang command package to clangCommon as it was breaking imports test.
